update listcustomer and list employee

+++ src/components/Customer/ListCustomerComponent.jsx
@@ -19,6 +19,8 @@ const ListEmployeeComponent = () => {
   const navigate = useNavigate();
   const accountId = localStorage.getItem("accountId");
   const [roleIdFilter, setRoleIdFilter] = useState('');
+  const [isAccountDropdownOpen, setAccountDropdownOpen] = useState(false);
+  const [showActiveModal, setShowActiveModal] = useState(false); 
 
   const filteredAccounts = accounts.filter(account => {
     const matchesRoleId = roleIdFilter ? account.roleId.includes(roleIdFilter) : true;
@@ -71,23 +73,41 @@ const ListEmployeeComponent = () => {
       });
   };
 
+  const addNewAccount = () => {
+    navigate('/add-account');
+  };
+
   const updateAccount = (accountId) => {
     navigate(`/edit-account/${accountId}`);
   };
 
-  const removeAccount = (accountId) => {
-    console.log(accountId)
-    deleteAccount(accountId)
-      .then(() => {
-        getAllAccounts();
-      })
-      .catch((error) => {
-        console.error("Error deleting employee: ", error);
-      });
-  };
+
 
   const handleSearchChange = (e) => {
     setSearchTerm(e.target.value);
+  };
+  const toggleAccountDropdown = () => {
+    setAccountDropdownOpen(!isAccountDropdownOpen);
+  };
+
+  const handleOpenActiveModal = () => {
+    setShowActiveModal(true);
+    getAllAccounts(); 
+  };
+
+  const handleDeactivateService = (accountId)=>{
+    if(accountId){
+      deleteAccount(accountId)
+      .then(()=>{
+        getAllAccounts();
+        
+      })
+      .catch((error) => console.error("Error creating service:", error));
+    }
+  }
+
+  const handleCloseActiveModal = () => {
+    setShowActiveModal(false);
   };
 
 
@@ -116,16 +136,43 @@ const ListEmployeeComponent = () => {
         
         <h6>List</h6>
         <li>
-          <a href="/manager"><i className="bi bi-person-badge me-2"><HiOutlineClipboardDocumentList /></i>Dashboard</a>
+          <a href="/manager"><i className="bi bi-person-badge me-2"><HiOutlineClipboardDocumentList /></i>Dashboard</a> 
         </li>
 
+        
+        <li onClick={toggleAccountDropdown} style={{ cursor: 'pointer', display: 'flex', alignItems: 'center' }}>
+             < a href="#"><i className="bi bi-person-badge me-2"><FiUsers /></i> Customers</a>
+            </li>
+            {isAccountDropdownOpen && (
+              <ul className="list-unstyled ms-3">
+                 <li>
+                 
+                 <a href="#" onClick={addNewAccount}>
+                   <i className="bi bi-person-badge me-2"> <IoMdAddCircle  /> </i> New Accounts
+                 </a>
+                </li>
+             
+               <li>
+              
+                 <a href="#" onClick={handleOpenActiveModal}>
+                   <i className="bi bi-person-badge me-2"> <IoMdAddCircle  /> </i> Deactive
+                 </a>
+               </li>
+             
+              </ul>
+            )}
         <li>
           <a href="/accounts"><i className="bi bi-person-badge me-2"><FiUsers /></i> Employees</a>
         </li>
 
+
+        
          <li>
             <a href="/service"><i className="bi bi-person-badge me-2"><HiOutlineClipboardDocumentList /></i> Services</a>
           </li>
+
+          
+      
         
         <h6>General</h6>
         <li>
@@ -143,6 +190,8 @@ const ListEmployeeComponent = () => {
         <li>
           <a href="#"><i className="bi bi-gear me-2"><IoSettingsOutline /></i> Settings</a>
          </li>
+
+      
        
       </ul>
       </nav>
@@ -150,7 +199,7 @@ const ListEmployeeComponent = () => {
       </aside>
       <main className="dashboard col-10 ">
       <header className="d-flex justify-content-between align-items-center mb-4 border-bottom pb-2 ">
-        <h4 className="title">Customers</h4>
+        <h4 className="title">Employees</h4>
             <header className="d-flex justify-content-between align-items-center mb-4 ">
            
             <div className="search-bar ml-auto">
@@ -191,11 +240,20 @@ const ListEmployeeComponent = () => {
        </div>
        <section className="delivery-ongoing-delivery mt-4 d-flex pt-3">
         <div className="delivery-list col-12">
-      
+          <div className="filter-bar d-flex mb-4 col-10 align-items-center">
+            <select className="form-select me-2 col-4" value={roleIdFilter} onChange={(e) => setRoleIdFilter(e.target.value)}>
+              <option value="">All Employees</option>
+              <option value="Sales">Sales</option>
+              <option value="Delivery">Delivery</option>
+            </select>
+            
+          </div>
+
           <table className="table table-striped table-bordered">
             <thead>
               <tr>
                 <th>Account Id</th>
+                
                 <th>First Name</th>
                 <th>Last Name</th>
                 <th>Email</th>
@@ -203,32 +261,93 @@ const ListEmployeeComponent = () => {
               </tr>
             </thead>
             <tbody>
-              {filteredAccounts.length > 0 ? (
-                filteredAccounts.map(account => (
-                  <tr key={account.accountId}>
-                    <td>{account.accountId}</td>
-                    <td>{account.firstName}</td>
-                    <td>{account.lastName}</td>
-                    <td>{account.email}</td>
-                    <td>
-                      <button className="btn btn-info" onClick={() => updateAccount(account.accountId)}>Update</button>
-                      <button className="btn btn-danger" onClick={() => removeAccount(account.accountId)}>Delete</button>
-                    </td>
-                  </tr>
-                ))
-              ) : (
-                <tr>
-                  <td colSpan="6" className="text-center">No Employees Found</td>
+            {filteredAccounts.filter(account => account.status === 1).length > 0 ? (
+              filteredAccounts.filter(account => account.status === 1).map(account => (
+                <tr key={account.accountId}>
+                  <td>{account.accountId}</td>
+                  
+                  <td>{account.firstName}</td>
+                  <td>{account.lastName}</td>
+                  <td>{account.email}</td>
+                  <td>
+                    <button className="btn btn-info" onClick={() => updateAccount(account.accountId)}>Update</button>
+                  </td>
                 </tr>
-              )}
-            </tbody>
+              ))
+            ) : (
+              <tr>
+                <td colSpan="6" className="text-center">No Employees Found</td>
+              </tr>
+            )}
+          </tbody>
           </table>
         </div>
       </section>
+
+      {showActiveModal && (
+      <div className="modal show d-block" role="dialog">
+        <div className="modal-dialog" role="document">
+          <div className="modal-content">
+            <div className="modal-header">
+              <h5 className="modal-title">Active Accounts</h5>
+              <button 
+                type="button" 
+                className="close" 
+                onClick={handleCloseActiveModal} 
+                aria-label="Close" 
+                style={{ marginLeft: '300px', backgroundColor: 'whitesmoke', cursor: 'pointer' }}
+              >
+                <span aria-hidden="true" style={{ fontSize: '20px', color: 'black' }}>&times;</span>
+              </button>
+            </div>
+            <div className="modal-body">
+              <table className="table table-striped">
+                <thead>
+                  <tr>
+                   
+                    <th>First Name</th>
+                    <th>Last Name</th>
+                    <th>Actions</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {accounts.map((account) => (
+                    <tr key={account.accountId}>
+                      
+                      <td>{account.firstName}</td>
+                      <td>{account.lastName}</td>
+                      <td>
+                        {account.status === 1 ? ( 
+                          <button
+                            className="btn btn-danger btn-sm"
+                            onClick={() => handleDeactivateService(account.accountId)}
+                          >
+                            Deactivate
+                          </button>
+                        ) : ( 
+                          <button
+                            className="btn btn-secondary btn-sm" 
+                            disabled 
+                          >
+                            Deactivate
+                          </button>
+                        )}
+                      </td>
+                    </tr>
+                  ))}
+                </tbody>
+              </table>
+            </div>
+            <div className="modal-footer">
+              <button type="button" className="btn btn-secondary" onClick={handleCloseActiveModal}>Close</button>
+            </div>
+          </div>
+        </div>
+      </div>
+    )}
       </main>
       </div>
     </div>
   );
 };
-
 export default ListEmployeeComponent;
